==> research/stage_02_data_transformation.ipynb <==
import os

os.chdir('../')

from dataclasses import dataclass
from pathlib import Path
@dataclass(frozen=True)
class Data_Transformation_Config:
    root_dir:  Path
    data_path: Path
    train_path:Path
    test_path: Path

from src.brain_tumor_classifier.constant import *
def create_directories(file_path):
    pass
def read_yaml(file_path):
    pass
class Configuration_Manager:
    def __init__(self,config_path=CONFIG_PATH,params_path=PARAMS_PATH,schema_path=SCHEMA_PATH):
        self.config = read_yaml(config_path)
        self.config = read_yaml(config_path)
        self.config = read_yaml(config_path)
        create_directories([self.config.artifacts_root])
        
    def get_data_transformation_config(self)->Data_Transformation_Config:
        config = self.config.Data_Transformation
        create_directories([config.root_dir])
        data_transformation_config = Data_Transformation_Config(root_dir=config.root_dir,data_path=config.data_path,train_path=config.train_path,test_path=config.test_path)
        return data_transformation_config
        

import os
import urllib.request as request
import zipfile
from src.brain_tumor_classifier.utils.common import logger,save_dataset
from src.brain_tumor_classifier.utils import logger
from tensorflow.keras.preprocessing.image import ImageDataGenerator,load_img
import numpy as np
import pandas as pd

class Data_Transformation:
    def __init__(self,config:Data_Transformation_Config):
        self.config = config
    
    def initiate_data_transformation(self):
        # Read Dataset
        IMAGE_SIZE = (224,224)
        train_data_generator = ImageDataGenerator(rescale=1./255,shear_range=0.2,zoom_range=0.2,horizontal_flip=True)
        test_data_generator = ImageDataGenerator(rescale=1./255)
        training_set = train_data_generator.flow_from_directory(os.path.join(self.config.data_path,'train'),target_size=IMAGE_SIZE,batch_size=32,class_mode='categorical')
        testing_set = test_data_generator.flow_from_directory(os.path.join(self.config.data_path,'test'),target_size=IMAGE_SIZE,batch_size=32,class_mode='categorical')
        save_dataset(training_set,self.config.train_path)
        save_dataset(testing_set,self.config.test_path)
        logger.info(f'Successfully applied Transformation techniques on data, and saved to: {self.config.root_dir}')
        return (training_set,testing_set)
